Support near field objects
These are objects that needs to be very close to the camera, for
example a cockpit, making them nearer than the near clipping
plane. The workaround is to render them in a later render queue,
in the renderQueueStarted() callback, set the frustum an clear the
depth buffer, and in the renderQueueEnded() callback, reset everything
back to the original settings. The objects needs to have it's bounding
box large enough so that it falls within the original frustum to not
be culled.